+++ uni-stark/src/scripts/bf_unistark.rs
@@ -16,6 +16,8 @@ use scripts::u31_lib::{
     u31ext_sub, u31ext_sub_u31, BabyBear4, BabyBearU31,
 };
 
+use crate::{get_generator, value_exp_n_minus_one};
+
 define_pushable!();
 
 // * input:
@@ -62,6 +64,8 @@ pub fn compute_acc_numerator_all<Val: BfField, Challenge: BfField>(
     quotient_chunk_nums: usize,
 ) -> Script {
     script! {
+        //push babybeay_generator.inverse()
+        {64944062 as u32}
         {u31ext_mul_u31::<BabyBear4>()}   // stack: zeta/a
         for i in 0..quotient_chunk_nums{
             OP_4TOALTSTACK
@@ -92,8 +96,7 @@ pub fn compute_acc_numerator_script<Val: BfField, Challenge: BfField>(
         "compute_acc_numerator",
         compute_acc_numerator_all::<Val, Challenge>(degree_bits, generator, quotient_chunk_nums),
     );
-    si.add_input(Val::generator().try_inverse().unwrap())
-        .add_input(zeta);
+    si.add_input(zeta);
     for i in 0..quotient_chunk_nums {
         si.add_output(numerators[i]);
     }
@@ -119,18 +122,18 @@ fn compute_acc_denominator<Val: BfField, Challenge: BfField>(
     quotient_chunk_nums: usize,
 ) -> Script {
     script! {
-        OP_1                                       //the accmulator is 1 when intialization
+        OP_1                                                //the accmulator is 1 when intialization
         for _ in 0..(quotient_chunk_nums - 1) {
-            OP_TOALTSTACK     //alt: acc     s:i,  (S-1-j)
-            OP_DUP            //alt: acc     s:i,i,(S-1-j)
-            OP_TOALTSTACK     //alt: i, acc  s:i,  (S-1-j)
-            OP_ADD            //alt: i, acc  s:i + (S-1-j) = lookup_index
+            OP_TOALTSTACK                                   //alt: acc     s:i,  (S-1-j)
+            OP_DUP                                          //alt: acc     s:i,i,(S-1-j)
+            OP_TOALTSTACK                                   //alt: i, acc  s:i,  (S-1-j)
+            OP_ADD                                          //alt: i, acc  s:i + (S-1-j) = lookup_index
             {get_generator::<Val>(generator, quotient_chunk_nums)}
-            {value_exp_n_minus_one::<Val>(log_n_conf)}     //alt: i, acc  s: (w^(i-j))^n - 1
+            {value_exp_n_minus_one::<Val>(log_n_conf)}      //alt: i, acc  s: (w^(i-j))^n - 1
             OP_FROMALTSTACK
             OP_SWAP
-            OP_FROMALTSTACK                           // s: acc, (w^(i-j))^n - 1, i
-            {u31_mul::<BabyBearU31>()}           // s: new_acc i j
+            OP_FROMALTSTACK                                 // s: acc, (w^(i-j))^n - 1, i
+            {u31_mul::<BabyBearU31>()}                      // s: new_acc i j
         }
         OP_TOALTSTACK
         OP_DROP  //DROP i
@@ -247,6 +250,8 @@ pub fn compute_quotient_zeta<Val: BfField, Challenge: BfField>(
     }
 }
 
+//SC::Challenge::monomial(e_i) * c
+//quotient value is EF, but we use EF as F_slice for commit.
 pub fn compute_quotient_i<Val: BfField, Challenge: BfField>() -> Script {
     script! {
         OP_0 OP_0 OP_0 OP_1
@@ -386,88 +391,10 @@ pub fn verify_quotient_script<Val: BfField, Challenge: BfField>(
     si
 }
 
-pub fn value_square_with_input<F: BfField>() -> Script {
-    script! {
-        if F::U32_SIZE == 1 {
-            OP_DUP
-            {u31_mul::<BabyBearU31>()}
-        }else{
-            OP_4DUP
-            {u31ext_mul::<BabyBear4>()}
-        }
-    }
-}
-
-pub fn value_exp_n_minus_one<F: BfField>(log_n: usize) -> Script {
-    script! {
-        for _ in 0..log_n {
-            {value_square_with_input::<F>()}
-        }
-        if F::U32_SIZE == 1 {
-            OP_1
-            {u31_sub::<BabyBearU31>()}
-        }else{
-            OP_0 OP_0 OP_0 OP_1
-            {u31ext_sub::<BabyBear4>()}
-        }
-    }
-}
-
-fn push_generator_table<F: BfField>(generator: F, quotient_chunk_nums: usize) -> Script {
-    script! {
-        for i in (0..quotient_chunk_nums).rev() {
-            for j in (0..F::U32_SIZE).rev(){
-                {generator.exp_u64(i as u64).as_u32_vec()[j]}
-            }
-        }
-        for i in 1..quotient_chunk_nums {
-            for j in (0..F::U32_SIZE).rev(){
-                {generator.exp_u64(i as u64).try_inverse().unwrap().as_u32_vec()[j]}
-            }
-        }
-    }
-}
-
-fn pop_generator_table<F: BfField>(quotient_chunk_nums: usize) -> Script {
-    script! {
-        for i in 0..(2 * quotient_chunk_nums - 1){
-            for j in (0..F::U32_SIZE).rev(){
-                OP_DROP
-            }
-        }
-    }
-}
-
-// stack: index  input: generator: w, quotient_chunk_nums: s
-fn get_generator<F: BfField>(generator: F, quotient_chunk_nums: usize) -> Script {
-    script! {
-        OP_TOALTSTACK
-        {push_generator_table::<F>(generator, quotient_chunk_nums)}
-        OP_FROMALTSTACK
-        if F::U32_SIZE == 1{
-            OP_PICK
-            OP_TOALTSTACK
-        }else{
-            OP_4MUL
-            OP_4PICK
-            OP_4TOALTSTACK
-        }
-
-        {pop_generator_table::<F>(quotient_chunk_nums)}
-
-        if F::U32_SIZE ==1{
-            OP_FROMALTSTACK
-        }else{
-            OP_4FROMALTSTACK
-        }
-    }
-}
-
 #[cfg(test)]
 mod tests {
-    use bitcoin::opcodes::{OP_FROMALTSTACK, OP_TRUE};
-    use bitcoin::ScriptBuf as Script;
-    use bitcoin_script::{define_pushable, script};
+    use bitcoin::opcodes::OP_TRUE;
+    use bitcoin_script::define_pushable;
     use fri::{FriConfig, TwoAdicFriPcs};
     use itertools::{izip, Itertools};
     use p3_commit::{PolynomialSpace, TwoAdicMultiplicativeCoset};
@@ -662,85 +589,4 @@ mod tests {
         );
         assert!(!res.success);
     }
-    // #[test]
-    // fn test_zps() {
-    //     type Challenge = BinomialExtensionField<BabyBear, 4>;
-    //     type Val = BabyBear;
-
-    //     let degree = 8; //n
-    //     let log_degree = log2_strict_usize(degree);
-    //     let quotient_degree = 4; //s
-    //     let log_quotient_degree = log2_strict_usize(quotient_degree);
-
-    //     let mut rng = ChaCha20Rng::seed_from_u64(0u64);
-
-    //     let zeta = rng.gen::<Challenge>();
-
-    //     let a = Val::generator().try_inverse().unwrap();
-
-    //     // let trace_domain = pcs.natural_domain_for_degree(degree);
-    //     let trace_domain = TwoAdicMultiplicativeCoset {
-    //         log_n: log_degree,
-    //         shift: Val::one(),
-    //     };
-
-    //     let quotient_domain =
-    //         trace_domain.create_disjoint_domain(1 << (log_degree + log_quotient_degree));
-
-    //     let generator = Val::two_adic_generator(quotient_domain.log_n);
-
-    //     let quotient_chunks_domains = quotient_domain.split_domains(quotient_degree);
-
-    //     let quotient_chunk_nums = quotient_chunks_domains.len();
-
-    //     let zps = quotient_chunks_domains
-    //         .iter()
-    //         .enumerate()
-    //         .map(|(i, domain)| {
-    //             quotient_chunks_domains
-    //                 .iter()
-    //                 .enumerate()
-    //                 .filter(|(j, _)| *j != i)
-    //                 .map(|(_, other_domain)| {
-    //                     other_domain.zp_at_point(zeta)
-    //                         * other_domain.zp_at_point(domain.first_point()).inverse()
-    //                 })
-    //                 .product::<Challenge>()
-    //         })
-    //         .collect_vec();
-
-    //     // let test_script = script! {
-    //     //     for i in (0..zps.len()).rev() {
-    //     //         {zps[i].as_u32_vec()[3]}{zps[i].as_u32_vec()[2]}{zps[i].as_u32_vec()[1]}{zps[i].as_u32_vec()[0]}
-    //     //     }
-    //     //     {zeta.as_u32_vec()[3]}{zeta.as_u32_vec()[2]}{zeta.as_u32_vec()[1]}{zeta.as_u32_vec()[0]}
-    //     //     {a.as_u32_vec()[0]}
-    //     //     {verify_quotient::<Val, Challenge>(
-    //     //         quotient_chunk_nums,
-    //     //         generator,
-    //     //         log2_strict_usize(degree))}
-    //     // };
-
-    //     // let res = execute_script(test_script);
-    //     // if !res.success {
-    //     //     println!("{:?}", res);
-    //     // }
-    //     // assert!(res.success);
-    //     let mut bc_assigner = DefaultBCAssignment::new();
-    //     let mut exec_script_info = verify_quotient_script::<Val, Challenge>(
-    //         zeta,
-    //         quotient_chunk_nums,
-    //         degree,
-    //         generator,
-    //         zps,
-    //     );
-
-    //     exec_script_info.gen(&mut bc_assigner);
-
-    //     let res = execute_script_with_inputs(
-    //         exec_script_info.get_eq_script(),
-    //         exec_script_info.witness(),
-    //     );
-    //     assert!(res.success);
-    // }
 }
