+++ uni-stark/src/scripts/mod.rs
@@ -1,3 +1,5 @@
 mod bf_unistark;
+mod utils;
 
 pub use bf_unistark::*;
+pub use utils::*;
